requirements.txt
# Core framework
fastapi
uvicorn[standard]
python-dotenv
pydantic

# Vector search (Restore FAISS)
faiss-cpu
numpy
openai

# Document processing
PyMuPDF

# Testing
pytest
httpx
